git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_NASA_challenge@2164 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixPilot/parameter_table.c
@@ -14,29 +14,29 @@ extern struct ADchannel udb_zaccel ;
 extern struct ADchannel udb_xrate ;
 extern struct ADchannel udb_yrate ;
 extern struct ADchannel udb_zrate ;
-extern int height_target_min ;
-extern int height_target_max ;
-extern int height_margin ;
+extern int16_t height_target_min ;
+extern int16_t height_target_max ;
+extern int16_t height_margin ;
 extern fractional alt_hold_throttle_min ;
 extern fractional alt_hold_throttle_max ;
-extern int alt_hold_pitch_min ;
-extern int alt_hold_pitch_max ;
-extern int alt_hold_pitch_high ;
-extern int rtl_pitch_down ;
-extern int minimum_groundspeed ;
-extern int maximum_airspeed ;
-extern int minimum_airspeed ;
-extern int desiredSpeed ;
-extern int minimum_groundspeed ;
-extern int maximum_airspeed ;
-extern int minimum_airspeed ;
-extern int cruise_airspeed ;
-extern int desiredSpeed ;
-extern int airspeed_pitch_min_aspd ;
-extern int airspeed_pitch_max_aspd ;
-extern int airspeed_pitch_adjust_rate ;
+extern int16_t alt_hold_pitch_min ;
+extern int16_t alt_hold_pitch_max ;
+extern int16_t alt_hold_pitch_high ;
+extern int16_t rtl_pitch_down ;
+extern int16_t minimum_groundspeed ;
+extern int16_t maximum_airspeed ;
+extern int16_t minimum_airspeed ;
+extern int16_t desiredSpeed ;
+extern int16_t minimum_groundspeed ;
+extern int16_t maximum_airspeed ;
+extern int16_t minimum_airspeed ;
+extern int16_t cruise_airspeed ;
+extern int16_t desiredSpeed ;
+extern int16_t airspeed_pitch_min_aspd ;
+extern int16_t airspeed_pitch_max_aspd ;
+extern int16_t airspeed_pitch_adjust_rate ;
 extern fractional airspeed_pitch_ki ;
-extern int airspeed_pitch_ki_limit ;
+extern int16_t airspeed_pitch_ki_limit ;
 
 
 const mavlink_parameter_parser    mavlink_parameter_parsers[] = {
@@ -124,7 +124,7 @@ const mavlink_parameter mavlink_parameters_list[] = {
 
     };
 
-const int count_of_parameters_list = sizeof(mavlink_parameters_list) / sizeof(mavlink_parameter);
+const int16_t count_of_parameters_list = sizeof(mavlink_parameters_list) / sizeof(mavlink_parameter);
 
 
 #endif
